bug fixed

+++ app/src/main/java/com/example/smartcart/LastShoppingActivity.java
@@ -86,7 +86,7 @@ public class LastShoppingActivity extends AppCompatActivity implements recyclevi
                             for (int i = 0; i < my_items.length; i++) {
                                 String item[] = my_items[i].split(",");
                                 total += Integer.parseInt(item[4]);
-                                item tmpItem = new item(item[1], item[0], Integer.parseInt(item[3]), Integer.parseInt(item[2]));
+                                item tmpItem = new item(item[1], item[0], Double.parseDouble(item[3]), Integer.parseInt(item[2]));
                                 tmp_items.add(tmpItem);
                                                             }
                             list.add(new order(Integer.parseInt(tmp.getKey()), total ,date, address, tmp_items));
